requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg==3.3.3
python-dotenv==1.1.1
sqlparse==0.5.5
pillow==11.0.0
